Import 2.1.38pre1

+++ include/linux/inet.h
@@ -45,8 +45,8 @@
 #ifdef __KERNEL__
 
 extern void		inet_proto_init(struct net_proto *pro);
-extern char		*in_ntoa(unsigned long in);
-extern unsigned long	in_aton(const char *str);
+extern char		*in_ntoa(__u32 in);
+extern __u32		in_aton(const char *str);
 
 #endif
 #endif	/* _LINUX_INET_H */
